% Generated by roxygen2 (4.0.2): do not edit by hand
\name{adaptivefence}
\alias{adaptivefence}
\title{Adaptive Fence model selection}
\usage{
adaptivefence(mf, f, ms, d, lf, pf, bs, grid = 101, bandwidth)
}
\arguments{
\item{mf}{function for fitting the model}

\item{f}{formular of full model}

\item{ms}{list of formular of candidates models}

\item{d}{data}

\item{lf}{function for lack of fit (to minimize)}

\item{pf}{function for pick measuring (to minimize)}

\item{bs}{bootstrap sample}

\item{grid}{grid for c}

\item{bandwidth}{bandwidth for kernel smooth function}
}
\value{
list with whatever
}
\description{
Adaptive Fence model selection
}

