% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fence.glmer}
\alias{fence.glmer}
\title{Fence model selection (Generalized Linear Mixed Model)}
\usage{
fence.glmer(full, data, family = c("b", "p"), B = 100, grid = 101,
  fence = c("adaptive", "nonadaptive"), cn = NA, bandwidth = NA,
  cpus = 2)
}
\arguments{
\item{full}{formular of full model}

\item{data}{data}

\item{family}{binary("b") or poisson("p")}

\item{B}{number of bootstrap sample, parametric for lmer}

\item{grid}{grid for c}

\item{fence}{fence method to be used, (adaptive, nonadaptive, more to add)}

\item{cn}{cn for nonadaptive}

\item{bandwidth}{bandwidth for kernel smooth function}
}
\value{
list with whatever
}
\description{
Fence model selection (Generalized Linear Mixed Model)
}
\note{
more link to be added
}
\author{
Jianyang Zhao
}

