% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fence.glmmadmb}
\alias{fence.glmmadmb}
\title{Fence model selection (Generalized Linear Mixed Model by glmmadmb)}
\usage{
fence.glmmadmb(full, data, family = c("g", "n", "p"), zeroInflation = FALSE,
  B = 100, grid = 101, fence = c("adaptive", "nonadaptive"), cn = NA,
  bandwidth = NA)
}
\arguments{
\item{full}{formular of full model}

\item{data}{data}

\item{family}{gamma("g") or negative binomial("n")}

\item{B}{number of bootstrap sample, parametric for lmer}

\item{grid}{grid for c}

\item{fence}{fence method to be used, (adaptive, nonadaptive, more to add)}

\item{cn}{cn for nonadaptive}

\item{bandwidth}{bandwidth for kernel smooth function}
}
\value{
list with whatever
}
\description{
Fence model selection (Generalized Linear Mixed Model by glmmadmb)
}
\note{
glmmadmb has more families but slower, more link need to be added and may fail in adaptive
}

