% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fence.sae}
\alias{fence.sae}
\title{Fence model selection (Small Area Estmation)}
\usage{
fence.sae(full, data, B = 100, grid = 101, fence = c("adaptive",
  "nonadaptive"), cn = NA, method = c("F-H", "NER"), D = NA,
  REML = FALSE, bandwidth = NA, cpus = 2)
}
\arguments{
\item{full}{formular of full model}

\item{data}{data}

\item{B}{number of bootstrap sample, parametric for lmer}

\item{grid}{grid for c}

\item{fence}{fence method to be used, (adaptive, nonadaptive, more to add)}

\item{cn}{cn for nonadaptive}

\item{REML}{REML}

\item{bandwidth}{bandwidth for kernel smooth function}
}
\value{
list with whatever
}
\description{
Fence model selection (Small Area Estmation)
}

