% Generated by roxygen2 (4.0.1): do not edit by hand
\name{IF.GSA}
\alias{IF.GSA}
\title{Invisible Fence gene selection}
\usage{
IF.GSA(gene, response, genenames, genesets, minsize = 0, B = 100,
  limitboot = length(genesets))
}
\arguments{
\item{gene}{a matrix, each column for a subject}

\item{response}{a factor vector, each element for a subject}

\item{genename}{a character vector}

\item{geneset}{a character list}

\item{minsize}{a integer, only use geneset which size larger or equal to this}

\item{B}{a integer, number of bootstrap samples}
}
\value{
list with whatever
}
\description{
Invisible Fence gene selection
}

