% Generated by roxygen2 (4.0.2): do not edit by hand
\name{IF.lmer}
\alias{IF.lmer}
\title{Invisible Fence model selection (Linear Mixed Model)}
\usage{
IF.lmer(full, data, B = 100, REML = TRUE, method = c("marginal",
  "conditional"))
}
\arguments{
\item{full}{formular of full model}

\item{data}{data}

\item{B}{number of bootstrap sample, parametric for lmer}

\item{REML}{REML}

\item{grid}{grid for c}

\item{fence}{fence method to be used, (adaptive, nonadaptive, more to add)}

\item{cn}{cn for nonadaptive}

\item{bandwidth}{bandwidth for kernel smooth function}
}
\value{
list with whatever
}
\description{
Invisible Fence model selection (Linear Mixed Model)
}

