% Generated by roxygen2 (4.0.2): do not edit by hand
\name{IF.nb}
\alias{IF.nb}
\title{Invisible Fence model selection (Linear Mixed Model)}
\usage{
IF.nb(full, data, B = 100, cpus = 2)
}
\arguments{
\item{full}{formular of full model}

\item{data}{data}

\item{B}{number of bootstrap sample, parametric for lmer}

\item{grid}{grid for c}

\item{fence}{fence method to be used, (adaptive, nonadaptive, more to add)}

\item{cn}{cn for nonadaptive}

\item{REML}{REML}

\item{bandwidth}{bandwidth for kernel smooth function}
}
\value{
list with whatever
}
\description{
Invisible Fence model selection (Linear Mixed Model)
}

