% Generated by roxygen2 (4.0.2): do not edit by hand
\name{invisiblefence}
\alias{invisiblefence}
\title{Invisible Fence model selection}
\usage{
invisiblefence(mf, f, d, lf, bs)
}
\arguments{
\item{mf}{function for fitting the model}

\item{f}{formular of full model}

\item{d}{data}

\item{lf}{a function provides lack of fit for all terms, e.g. function(x) abs(coef(x))[-1]}

\item{bs}{bootstrap sample}
}
\value{
list with whatever
}
\description{
Invisible Fence model selection
}

