% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nonadaptivefence}
\alias{nonadaptivefence}
\title{Nonadaptive Fence model selection}
\usage{
nonadaptivefence(mf, f, ms, d, lf, pf, cn)
}
\arguments{
\item{mf}{function for fitting the model}

\item{f}{formular of full model}

\item{ms}{list of formular of candidates models}

\item{d}{data}

\item{lf}{function for lack of fit (to minimize)}

\item{pf}{function for pick measuring (to minimize)}
}
\value{
list with whatever
}
\description{
Nonadaptive Fence model selection
}

