% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodAF.R
\name{plot.AF}
\alias{plot.AF}
\title{Plot Adaptive Fence model selection}
\usage{
\method{plot}{AF}(x = res, ...)
}
\arguments{
\item{x}{Object to be plotted}

\item{...}{Additional arguments. CNot currently used.}
}
\description{
Plot Adaptive Fence model selection
}
