% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodNF.R
\name{plot.NF}
\alias{plot.NF}
\title{Plot Nonparametric Fence model selection}
\usage{
\method{plot}{NF}(x = res, ...)
}
\arguments{
\item{x}{Object to be plotted}

\item{...}{Additional arguments. CNot currently used.}
}
\description{
Plot Nonparametric Fence model selection
}
