% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodAF.R
\name{summary.AF}
\alias{summary.AF}
\title{Summary Adaptive Fence model selection}
\usage{
\method{summary}{AF}(object = res, ...)
}
\arguments{
\item{object}{Object to be summarized}

\item{...}{addition arguments. Not currently used}
}
\description{
Summary Adaptive Fence model selection
}
