% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodNF.R
\name{summary.NF}
\alias{summary.NF}
\title{Summary Nonparametric Fence model selection}
\usage{
\method{summary}{NF}(object = res, ...)
}
\arguments{
\item{object}{Object to be summarized}

\item{...}{addition arguments. Not currently used}
}
\description{
Summary Nonparametric Fence model selection
}
