% Generated by roxygen2 (4.0.2): do not edit by hand
\name{NF}
\alias{NF}
\title{Nonparametric Fence}
\usage{
NF(full, data, spline, ps = 1:4, qs = NA, B = 100, grid = 101,
  bandwidth = NA)
}
\arguments{
\item{full}{formular of full model}

\item{data}{data}

\item{spline}{variable need spline terms}

\item{ps}{order of power}

\item{qs}{number of knots}

\item{B}{number of bootstrap sample, parametric for lmer}

\item{grid}{grid for c}

\item{bandwidth}{bandwidth for kernel smooth function}
}
\value{
list with whatever
}
\description{
Nonparametric Fence
}

